requirements.txt
asgiref==3.5.2
Django==4.0.4
django-cors-headers==3.12.0
djangorestframework==3.13.1
django-base64field==1.0
djongo==1.3.6
dnspython==2.2.1
pymongo==3.12.1
pytz==2022.1
sqlparse==0.2.4
tzdata==2022.1
gunicorn==20.1.0
whitenoise==6.1.0